Config cleaned up
