+++ extensions/kotlin/src/main/kotlin/org/modelmapper/KModelMapper.kt
@@ -1,0 +1,9 @@
+package org.modelmapper
+
+import org.modelmapper.kotlin.KotlinConstructorInjector
+
+class KModelMapper : ModelMapper {
+    constructor():super(){
+        configuration.setConstructorInjector(KotlinConstructorInjector())
+    }
+}
